feature : 요청에서 전송받는 token handling 구현(1)
