+++ api/src/main/java/com/bookwhale/config/WebMvcConfig.java
@@ -1,0 +1,39 @@
+package com.bookwhale.config;
+
+import com.bookwhale.auth.service.OauthService;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebMvcConfig implements WebMvcConfigurer {
+
+    private static final long MAX_AGE_SECS = 3600;
+
+    private final OauthService oauthService;
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+            .allowedOriginPatterns("*")
+            .allowedMethods("GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS")
+            .allowedHeaders("*")
+            .allowCredentials(true)
+            .maxAge(MAX_AGE_SECS);
+    }
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
+        argumentResolvers.add(createAuthenticationArgumentResolver());
+    }
+
+    @Bean
+    public CurrentUserArgumentResolver createAuthenticationArgumentResolver() {
+        return new CurrentUserArgumentResolver(oauthService);
+    }
+}
